requirements.txt
asyncio
pyyaml
psycopg2-binary
ph4_walkingpad